updated preprocess_SES.py
version with .astype float 16 removed

+++ Step6_SensitivityAnalyses/SES/preprocess_SES.py
@@ -46,7 +46,7 @@ SESfeatures = pd.read_csv('/gpfs/fs001/cbica/projects/abcdfnets/SES_features.csv
 SESfeaturesOrdered=SESfeatures[SESfeatures.subjectkey==phenotypes.subjectkey.values]['reshist_addr1_adi_income']
 features=np.array(SESfeaturesOrdered)
 features = features.reshape(features.shape[0],-1) #subject by flat array shape to be flat for sklearn
-np.save('/{0}/features_{1}.npy'.format(outdir,network),features.astype(np.float16))
+np.save('/{0}/features_{1}.npy'.format(outdir,network),features)
